# app/Http/Requests/StoreBookRequest.php
<?php

    namespace App\Http\Requests;

    use Illuminate\Foundation\Http\FormRequest;

    class StoreBookRequest extends FormRequest
    {

        public function rules (): array
        {
            return [
                "author_id" => "required|integer",
                "title" => "required|string",
                "overview" => "required|string",
                "isbn" => "required|string",
                "publisher" => "required|string",
                "publication_date" => "required|string",
                "edition_description" => "require|string",
                "pages" => "require|integer",
                "lexile" => "require|string",
                "sales_rank" => "nullable|integer",
            ];
        }

        public function messages ()
        {
            return parent::messages(); // TODO: Change the autogenerated stub
        }
    }